KV Tables

+++ sdk/tables/azure-data-tables/test/ut/table_client_test.cpp
@@ -140,6 +140,11 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, CreateTable)
   {
+    if (GetParam() != AuthType::SAS)
+    {
+      SkipTest();
+      return;
+    }
     auto createResponse = m_tableServiceClient->CreateTable(m_tableName);
     EXPECT_EQ(createResponse.Value.TableName, m_tableName);
     EXPECT_EQ(createResponse.Value.EditLink, "Tables('" + m_tableName + "')");
@@ -198,7 +203,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, ListTables)
   {
-    if (GetParam() == AuthType::ConnectionString)
+    if (GetParam() == AuthType::ConnectionString || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
     }
@@ -225,6 +230,11 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, DeleteTable)
   {
+    if (GetParam() == AuthType::SAS)
+    {
+      EXPECT_TRUE(true);
+      return;
+    }
     auto createResponse = m_tableServiceClient->CreateTable(m_tableName);
 
     auto response = m_tableServiceClient->DeleteTable(m_tableName);
@@ -238,7 +248,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, ServiceClientGetProperties)
   {
-    if (GetParam() == AuthType::ConnectionString)
+    if (GetParam() == AuthType::ConnectionString || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
     }
@@ -278,7 +288,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityCreate)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -298,7 +308,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityUpdate)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -332,7 +342,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityMerge)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -366,7 +376,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityDelete)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -401,7 +411,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityUpsert)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -440,7 +450,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityQuery)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
@@ -481,7 +491,7 @@ namespace Azure { namespace Data { namespace Test {
 
   TEST_P(TablesClientTest, EntityGet)
   {
-    if (GetParam() == AuthType::Key)
+    if (GetParam() == AuthType::Key || GetParam() == AuthType::SAS)
     {
       EXPECT_TRUE(true);
       return;
